requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
Django==3.1.2
django-celery-results==1.2.1
django-redis==4.12.1
djangorestframework==3.12.1
kombu==4.6.11
psycopg2==2.8.6
pytz==2020.1
redis==3.5.3
sqlparse==0.4.1
vine==1.3.0
